+++ src/senaite/core/interfaces/__init__.py
@@ -455,3 +455,26 @@ class IDynamicAnalysisSpec(Interface):
 class IDynamicAnalysisSpecs(Interface):
     """Marker interface for Dynamic analysis specs folder
     """
+
+
+class ILabProduct(Interface):
+    """Marker interface for lab product
+    """
+
+
+class ILabProducts(Interface):
+    """Marker interface for lab products folder
+    """
+
+
+class IHavePrice(Interface):
+    """Marker interface for objects that have a Price
+    """
+
+    def getPrice(self):
+        """Returns the price of the instance
+        """
+
+    def getTotalPrice(self):
+        """Returns the total price of the instance
+        """