+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 2.6.0 (unreleased)
 ------------------
 
+- #2581 Migrate Suppliers to Dexterity
+- #2574 Migrate LabProducts to Dexterity
 - #2583 Make SampleContainer DX type inherit from core's DX Container
 - #2582 Do not override dependent values with empties
 - #2573 Move Interpretation Templates to Senaite setup folder